Defect FW-1951: very large volumes overran int size (long should have been used).
Former-commit-id: 46a97936b583536fb4378a4edc05bf0cd64d5b34

+++ console/src/main/java/org/janelia/it/FlyWorkstation/gui/alignment_board_viewer/volume_builder/ByteChannelDelegate.java
@@ -27,7 +27,7 @@ public class ByteChannelDelegate {
             ChannelMetaData srcChannelMetaData,
             ChannelMetaData targetChannelMetaData,
             byte[] channelData,
-            int targetPos
+            long targetPos
     ) {
         int[] orderedRgbIndexes = srcChannelMetaData.getOrderedRgbIndexes();
         byte[] targetChannelBytes = new byte[ targetChannelMetaData.byteCount * targetChannelMetaData.channelCount ];
@@ -69,7 +69,7 @@ public class ByteChannelDelegate {
     public void padChannelBytes(
             ChannelMetaData srcChannelMetaData,
             ChannelMetaData targetChannelMetaData,
-            int targetPos
+            long targetPos
     ) {
         // Pad out to the end, to create the alpha byte.
         if ( targetChannelMetaData.channelCount >= ( srcChannelMetaData.channelCount + 1 ) ) {
