+++ console/src/main/java/org/janelia/it/FlyWorkstation/gui/alignment_board_viewer/volume_builder/FirstInToByteInterpreter.java
@@ -34,7 +34,7 @@ public class FirstInToByteInterpreter implements ChannelInterpreterI {
             ChannelMetaData targetChannelMetaData,
             int fileMaskNum,
             byte[] channelData,
-            int targetPos
+            long targetPos
     ) {
         boolean wroteBack = false;
